feat: extracted HTTPFetcher and Fetcher interface from LoanDataFetcher
this removes the dependency on HTTP stuff from LoanDataFetcher by injecting the Fetcher into LoanDataFetcher
also renamed LoanDataFetcher to FetchLoanDataUseCase since it's now definitively a use case
